An empty (no production) grammar, albeit illegal, should still compile...
... we compile it as a grammar that completes upon an immediate EOF input.

Technically it would be more accurate to create a grammar that always issues a
syntax error (since no input is acceptable, not even an empty input), but that
doesn't allow us to compile a table. The generated code should minimize exceptions
and prefer exercising the same code path; hence we do generate a grammar.

Note that this synth S is now an explicit non-terminal, and the EOF sym is now
an explicit terminal, so array allocations still work when the input grammar
is empty.