Generated _scan to _lex rename, generate a different _scan function..
What we previously called the _scan function (driving the scantables
based on input) is now the _lex function.

A new _scan function is generated which (work-in-progress) is to operate like
the _parse function, but take not a symbol but a text buffer as input.

Will want to change the function prototypes around to something that fits, not
final yet.