prd_grammar.c/h: remove _alt.c/h and add as regular C files...
... even though they're generated. This brings them inline with regex_grammar.cnd/c/h.

It may be desirable to generate these automatically from the .cnd, however, the
tail eating snake that creates in terms of build might be unstable (eg. cleaning
the build output would delete some of the source needed to build the executable
needed to build the source that builds the executable, etc. etc. bootstrap problem)