Support for %raii_constructor, and duplicate types now permissable...
... %raii_constructor changes the semantics of construction in that, if it
returns, it is deemed to have failed construction (eg. a constructor that
throws, for instance.) The regular %constructor has the opposite semantics
(if it returns, it is deemed to have succeeded.)

... duplicate types are now distinct types, where before they would have
been the same. Consequently, one int can now represent a file-descriptor and
have appropriate -1 constructor and close() destructor semantics, while the
other int may now represent an integer value, concurrently, with different
constructor and destructors assigned. Before, the "int" would have been
looked up and shared between the two. While writing the manual this behavior
was hard to justify.