Fixes on cleanup of top of stack symbol, and others..
.. _cleanup and _reset now don't issue continuation labels (negative value
caused issues.)

.. top_of_stack_has_common_data_ and top_of_stack_has_sym_data_...
these two maintain the constructor state of the top most symbol data. This is
necessary for the parse() function, where will fill the top of the stack
directly rather than via slot[0] or slot[1]. Consequently, there are three
states, not 1:
State 0: the slot has been allocated (push happened), but no common or symbol
data.
State 1: there is common data but no symbol data
State 2: there is both common and symbol data.

These states may occur when the parse() function moves in a token, when
constructors for common or symbol data return to the caller and _cleanup
is then called.