Rename %on_success to %on_finish and apply everywhere we return 0...
... this is more accurate as the 0 return value does not indicate success,
rather, it indicates we're done. This indication can come from a number
of places, not just where we reduce the synthetic S, but also where we
conclude that errors (already reported earlier) cannot be recovered from.

So, renamed to on_finish and applied the on_finish_snippet everywhere we
return 0. Therefore users can now fully capture the return value behavior
of the generated parsers.