New regular expression engine...
... aims to provide modes (start conditions where some patterns may match and
not others) and anchors (match non-symbols such as start of input, start of line,
end of input and end of line.)

Also, internally it is not tied to 256 input characters, but any range in uint32_t
can (internally) be processed. This is to set it up for later UTF-8/16/32 support.

Currently just emits a table that is compatible with its predecessor, more work
needed to emit the aforementioned features.