Modes feature...
... this allows patterns to be selectively applied depending on what mode the
scanner / lexer is in. Set the mode using <prefix>set_mode(), using one of the
M_<PREFIX>XXX constants, where XXX is either one of the modes declared using
the %mode directive or the default mode "default" (for which the constant
becomes M_<PREFIX>DEFAULT.)

To apply a mode to a pattern, first create a mode group:
<MODES> {
  .. patterns ..
}

Each of these patterns are only active when one of the MODES is selected.

MODES is a comma separated list of mode identifiers declared using %mode, or
alternatively the "default" mode identifier.

An alternative version where <MODE> <pattern> is used (without the curly
braces) is planned (and will parse) but not yet implemented.