Fix: error location on end of input
Work much harder to ensure that the tokenizer start_col_, start_line_ and
filename_ fields are populated when we are passing not input, but an
end-of-input marker. We do this by lingering the last location of the last
xlts processed by the tokenizer in these fields.

Note that for most error reporting, the matched XLTS will provide the
location information, this only applies where we issue an error without
any accompanying matched XLTS in the tokenizer.