Add $set_mode(), and $chg_term() ...
$set_mode(mode) allows changing the mode of the scanner when inside a pattern
action. mode is an identifier that matches up with a %mode
declaration (eg. case insensitive, and dashes are permitted.)

$chg_term(terminal) allows changing the matched terminal when inside a pattern
action. terminal is an identifier that matches up with a %token declaration
(eg. here too, case insensitive, and dashes are permitted.) Extreme care
should be taken when using this, its use-case is late resolution of scanner
ambiguities in some languages (eg. typedef-name in C99). The rest of the
scanner will blindly assume the matched terminal was always the new terminal
set by $chg_term() so there may be dangerous conditions arising where the
wrong objects are deconstructed, or not deconstructed, or garbled memory
is deemed already initialized.. The data for the old terminal and the data
for the new terminal set by $chg_term() are guaranteed to be at the same
address, this property can be used to ensure safety if the datatypes are
different, furthermore, there is no problem if the datatypes and constructor
/destructors are the same between the terminals. Needs careful use.