fix: incorrect resumption of scanning at buffer feedme..
The match_index would not be advanced causing the same piece of input to be
evaluated again; potentially leading to either a mismatch in scan_state and
the input (eg. lexical errors where there were none) or, in the case of
very small matches, an incorrect match of size 0, which, being 0, would
cause the lexer to never reset itself.