Fix: issue syntax errors when parsing grammar productions ...
Parser recovery upon erroneous input with no "error" token recovery productions
in the grammar incorrectly requested another token (thus ignoring the erroneous
token silently.)

The desired behavior is for it to return a syntax error or, if that syntax error
is temporarily muted (as per documented behavior), only then request a next token.

This affects users of the parser-only functionality <prefix>parse(). Users of
<prefix>scan() functionality should not be affected.