$is_mode(mode)
To check if the current mode is `mode`, use the `$is_mode(mode)` special
identifier. Analogous to `$set_mode()`, this will look up the `mode` identifier
(which can be represented in the same manner as in the `%mode` directive,
that is, case-insensitive and optionally using dashes (`-`)), and translate
it in the generated code to an inline comparison of the current mode to the
appropriate `M_<PREFIX>MODE` constant.

The value is 1 if the mode is the same, or 0 otherwise.

The `$is_mode(mode)` special identifier is available from any action.