fix: preserve whitespace in types rather than ...
... assuming the tokenization of the target language is identical to your
own. This fix stops inserting spaces between tokens in a type in favor of
preserving the whitespace as it was in the .cbrt input file.

A downside is that comments are seen as whitespace, so they too, will be
preserved. For our current target languages (C/C++) this is fine.

Practical issue being addressed here is that "::" is a scoping token,
which should be preserved but is seen as two separate colon ":" tokens,
which are being separated by the old code.

Reason the old code separated tokens is because "unsigned long" should not
become "unsignedlong". By preserving the whitespace, both behaviors are
supported at the same time, without having to introduce all target language
tokens to the carburetta tokenizer.