Support for %move directive
The %move directive allows better support for C++ classes (and builds us up
to an eventual %class C++ shorthand for the %type directive.)

%move is to be used alongside %constructor and %destructor directives, and
describes how to move the datatype from $0 to $$.

So, for instance:
%move $$ = $0
would use an assignment to move, whereas:
%move $$ = std::move($0)
would use C++ move semantics.

If no %move directive is specified for a type, then a memcpy is performed.

*NOTE*:
As part of this change, constructors and destructors are more strictly called,
so for any move, the destination's constructor is first called, then the move
is performed, and finally the source's destructor is called.
*This is different from before* and may impact code compatibility, previously,
a memcpy (possibly as part of a realloc()) was performed for a move operation,
and no constructors or destructors would have been called. Care needs to be
taken that destructors are not called twice; %move should be used to blank out
the source such that the destructor becomes harmless.