Support for %class directive
When using C++, rather than explicitly having to define a %raii_constructor,
%move, and %destructor for each and every %type, you can now use %class instead
of %type and have carburetta fill these in for you.

See for instance the t8 test case, where we now have:

  %class expr term factor value INTEGER: Value

whereas before we had:

  %type expr term factor value INTEGER: Value
  %raii_constructor new (&$$) Value();
  %move $$ = std::move($0);
  %destructor $$.~Value();

When encountering %class, the %raii_constructor, %move, and %destructor are
now generated implicitly. The code for this is slightly more involved than
one might imagine as "%class x: MyType $[4]" is still valid just as it was
with %type.

So the current code, using %class, is much shorter and more expressive &
friendly for newcomers.

Using %class causes C compilation to fail (because it makes the output
C++ specific.)