<prefix>token_common_data() function...
... in specific cases it can be very useful to get to the common data for
the current token. For instance, if _<PREFIX>SYNTAX_ERROR is returned, and
the common data is used to track the source location of tokens and symbols,
then getting to the common data of the token where the error occurred in an
easy non-roundabout way is valuable.

The <prefix>token_common_data() function does just that, returning a pointer
to the common data if there is a current token and if a %common_type or
%common_class definition exists. The "current token" is the token as it
exists prior to being shifted onto the LR stack. It should only be relied to
exist after _<PREFIX>SYNTAX_ERROR is returned, although the failure case
is NULL will be returned.

The type of the pointer returned is a void*, this may be irksome but ensures
that the common datatype is not exposed in any generated header (doing so
would require dependencies to be met prior to including the header, which
would potentially make Carburetta more cumbersome to use and break existing
code.)

Callers should check if NULL is returned, and if not, cast with confidence to
the appropriate type for %common_type or %common_class.