fix: corruption caused by non-destructor types.
when using %type (not %class) without an explicit %destructor specified, there
was corruption on breaks not being generated (and so it would fall through to
a handler for a different type; if a different handler followed it.)

Problem shows up when you have at least one %type with no destructor and there
are multiple distinct types. If %class is used, a destructor is always generated.