Change of default extern "C" behavior, %externc and %no_externc directives...
... When %class or %common_class is used, Carburetta knows that C++ code will be
generated. Previously, it still generated the extern "C" header guard, for
consistency's sake, but this creates extra work when compiling as C++: The
generated code file does not include the header file unless the user does so
explicitly, consequently, the code file emitted a decorated set of functions
whereas the header file specified a plain-C external linkage. Out of the box,
this would cause a link error and require new users to analyse and know what's
going on and explicitly include the header.

The new behavior is for Carburetta to not generate "extern "C"" guards when it
knows the code is C++ (when %class or %common_class is used at least once.)
Additionally, there are now two directives to override this behavior:

%externc : Generate extern "C" even if this is a C++ scanner/parser.
%no_externc : Do not generate extern "C", regardless of whether this is a C or
C++ scanner/parser.