Tilly example (WIP) ...
Tilly is a new example, under construction, for Carburetta.

It implements a Tile Matching algorithm (hence "Tilly", a cutesy derivation
from "Tile".)

The tile matching algorithm is from the '89 paper:
"Code Generation Using Tree Matching and Dynamic Programming"
Alfred V. Aho, Mahadevan Ganapathi, and Steven W.K. Tjiang.
https://doi.org/10.1145/69558.75700

The Twig language described therein is lost in time (I googled around for it
but couldn't find it.)

Tile matchers are useful for instruction selection for arbitrary architectures
from expression trees, so the example has some applicability in the realm of
compiler construction.

Committing it now, though it'll be a while before this example has reached
maturity. A work in progress...