fixes problematic interactions of absence of %constructor/%move/%destructor
fix 1: If, for a %type, there is no %constructor, no %move, and no %destructor
specified, then the generated code will not copy the data from the scanner
to the parser.

fix 2: If, for a %type, there is no explicit %constructor, but there is an
explicit %move, then carburetta will generate code to clear the destination data
to zero prior to issuing the %move, but that generated code would not compile.

C++ %class is not affected as %constructor, %move and %destructor are all
implicitly generated (and so exist).