fix: allow %% .. %% code snippets when inside %mode <MODE> { ... }
Should allow %% code snippet sections when inside a <MODE> { ... } mode
specifier for the scanner.

The actual problem here is that we "finished" (completed scanning and parsing)
of a section when we switched out to a code snippet using %%, and then re-started
scanning a section when we switched back from a code snippet using %% (doing so
implicitly.) Consequently, when a section closed using %%, the parser expected
a closing } on the mode specifiers or else the grammar would be incomplete.

The fix is to not "finish" scanner and grammar parsing until all input has
been exhausted, keeping the scanner and parser states for scanner and grammar
sections open until the end. The downside is that this now requires scanner and
grammar parsers to complete for all input configurations, even if an input
skips one of these sections. This works because both regex_grammar.cbrt and
prd_grammar.cbrt are well-behaved on empty input.