tilly should have one match word for each tile.
The idea being we maintain matching prefixes for each possible tile where
the distance is encoded by the number of the bit set. Now most tiles will be
small (not very deep branches), and so 64 bit positions would be an excessive
over-allocation, but for now this is a simple approach for getting to a proof
of concept.