tilly: Let "$$" refer to the root node of the pattern..
.. so eg. suppose we have:
modrm_ea: reg64 {
  computed_cost = $$->reduction_cost_[LABEL_REG64 - NUM_SYMS];
} = {}

then $$ refers to the reg64 node. You might think, following the analogue from
LR parsing, that this should be $0$ (the first argument) but then the 0 is a
dud because it is *always* the first argument as the right-side pattern is a
tree. The 0 being redundant, with $$ we can leave it out. This also makes
$.0$ and $0$ the same thing.