tilly fix: comments in pattern sections caused tile snippets to be off ...
... the mlc (multi-line comment) stage substitutes a /**/ comment to a single
space, so that directives can have comments and they can extend into the next
line. The dis (directive input splitter) stage routes to the appropriate stage
based on that substituted text (e.g. routes to tile processing or routes to
code prologue sections, and switches between these.) However, *after* it is
done with this, it should pass the text and situs source range data to the
tiles section *without* any substitutions (the tiles section will perform its
own scanning of comments and the like.) This we failed to do, we passed the
text data in without the substitutions, but passed the situs source range
data with the substitutions. Because of that, the tiles did not record the
appropriate source code ranges for their actions, causing us to mis-generate
the action whenever a comment appeared in a non-code section.