Direct cost attribute access and also ...
... supporting # and @ in pattern actions.

Direct cost attribute access means instead of this:
scaled_r64: mul(eight, r64) {
  computed_cost = $1$->reduction_cost_[LABEL_R64 - NUM_SYMS];
} = { ...stuff... }

we can now do this:
scaled_r64: mul(eight, r64) {
  computed_cost = $@1$;
} = { ...stuff... }

so the "$@1$" has a path that works like "$1$" but the @ causes it to access
the appropriate reduction_cost_ for the label at that path of the tile we're
reducing.

This requires supporting '@' in the scanners and while we're at it I also
added the '#' character.