tilly: Allow cyclic tile dependencies ..
.. so before, if you had:
a: b {
  cost = blah;
} = {
  ..
}

then you had to make sure that the tile pattern "a: b" would follow the tile
pattern for "b: .. whatever .." as otherwise it would be missed. Note that this
was not a problem for more complicated tiles (e.g. "a: b(c)"), only for the
simpler ones. Given that this is a scenario we'd like to support without
requiring too much thought when writing the tile patterns, we now keep looping
while tiles are being reduced, and double-check that we actually did anything
as a termination condition.