%visit and %visit_params directives (WIP) ...
... generate a <prefix>stack_visit() function for user custom book-keeping.
The idea here is that, inside the stack, exists runtime type information on
what type of symbols are inside the stack. For some use-cases (garbage
collection, debug, or other), it may be desirable to iterate over all user-
defined symbols by their custom type and execute a user-defined snippet,
allowing the user to leverage the rtti inside the stack.

To emit the <prefix>stack_visit(), specify a %visit snippet to execute for
at least one %type, %class, %common_type or %common_class. The %visit
directive should include the snippet to execute.

To pass custom arguments to the snippets, use the %visit_params directive,
much like you would the %params directive for the generated scan and parse
functions.

Additionally, fixed an error message to be more inclusive of the directives
that the various type related attribute directives are allowed to follow.