fix: correct #line number that returns us to the output location ..
.. the snippet location (the location within the .cbrt file) was correct,
but the output location (the location we'd like the compiler to return to
after emitting the snippet, the carburetta specific, non-user, code) was not
correct because, at the moment we generate that #line, we don't actually know
the current line number yet. Reason being that there is code at the top of
the function (where we dispatch to our continuation point in a large switch-
goto table) that is not generated until *after* the entire body of the
function has been generated.

This fix adapts out 'retained' indenting printer to retain the abstract notion
of a "line number + 1", to be settled when we perform the final output. Thus
the number is accurate at that point.