New --sym-names (short -n) commandline option ..
.. when specified, --sym-names will generate a symbol name table. This allows
you to take the integer symbol value (works for both a terminal or a non-
terminal, their ranges do not overlap), and use it to index the generated
"const char *<prefix>symbol_names_[]" table to get the name string for the
symbol. The name string is identical to the string you used to declare the
symbol in the carburetta input file, e.g. if you specify "%token My-Token"
then the name "My-Token" is returned, rather than the C-identifier "MY_TOKEN"
that carburetta translates it to.

.. As a safety measure, there is also an "int <prefix>symbol_names_length_"
variable that holds the maximum value of any symbol + 1, and so corresponds
to the length of the table. This should be used to range check any access on
the array for safety reasons. The type 'int' is used to match the type that
you pass symbols in with to the parse function in parser-only configurations.

.. Finally, note that there are possibly gaps in the table (where temporary
generated symbols exist internal to the algorithms of Carburetta) - such
entries have a NULL in them.

.. The default is off, because the string table can consume a bit of size,
and is likely not used frequently. You have to turn this on explicitly with
"--sym-names".