prologue, epilogue and header #line directives (WIP)...
... remove parts and use xlts infrastructure to manage the prologue, epilogue
and header section text. Also generate #line at the start of these sections
when we generate them.

Note that the line numbers are currently incorrect. Part of the explanation is
that the prologue in particular may consist of multiple distinct sections in
the source (and so will need more careful generation to handle the multiple
#line directives that are needed.) There is also an off by one I have not
yet accounted for on the epilogue. Work in progress.