fix: inaccurate accounting of line # of newlines themselves ...
... the desired behavior (see for instance comments in xlts.h) is for a newline
to be on the line it ends, not the line it starts. Actual behavior was for an
empty chunk to be inserted prior to the newline, for the line number to then
be advanced, and a chunk of 1 newline to be issued on the next line number.

This caused problems for cases where we were trying to locate the line number
of an empty line, because the 0-sized chunk would have been stripped, and the
newline would appear one line beyond where it should have been.

Through this fix, the #line directive of the epilogue now gives the correct
line number even if the epilogue starts with newlines.

Note however, that prologues in particular, may have "jumps" to account for
intermission code that is later appended to it. These intermissions are not
yet given the correct line number.